Feat:  weeklyPopular data fetching  && auth > base
