+++ src/api/post.ts
@@ -154,14 +154,8 @@ export const getTodayCoffeeInfo = async () => {
 // WeeklyPopular
 export const getWeeklyPopular = async () => {
   try {
-    setTimeout(deleteCachedData, 7 * 24 * 60 * 60 * 1000);
-    const res = await authInstance.get('/popular');
+    const res = await baseInstance.get('/popular');
     await useSetCacheData('brand', '/WeeklyPopular', res.data.data);
-    function deleteCachedData() {
-      caches.open('brand').then(cache => {
-        cache.delete('/popular');
-      });
-    }
     return res.data.data;
   } catch (error) {
     console.log('Failed to get Weekly Popular List', error);
